move types completely to v1
