--- node/core/approval-voting/src/approval-db/v1.rs
@@ -1,109 +1,0 @@
-// Copyright 2020 Parity Technologies (UK) Ltd.
-// This file is part of Polkadot.
-
-// Polkadot is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Polkadot is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
-
-//! Version 1 of the DB schema.
-
-use bitvec::{vec::BitVec, order::Lsb0 as BitOrderLsb0};
-use polkadot_primitives::v1::{
-	ValidatorIndex, GroupIndex, CandidateReceipt, SessionIndex, CoreIndex,
-	BlockNumber, Hash, CandidateHash,
-};
-use polkadot_node_primitives::approval::{DelayTranche, AssignmentCert};
-use sp_consensus_slots::Slot;
-
-// slot_duration * 2 + DelayTranche gives the number of delay tranches since the
-// unix epoch.
-pub type Tick = u64;
-
-pub type Bitfield = BitVec<BitOrderLsb0, u8>;
-
-/// Details pertaining to our assignment on a block.
-#[derive(Encode, Decode)]
-pub struct OurAssignment {
-	pub cert: AssignmentCert,
-	pub tranche: DelayTranche,
-	pub validator_index: ValidatorIndex,
-	// Whether the assignment has been triggered already.
-	pub triggered: bool,
-}
-
-/// Metadata regarding a specific tranche of assignments for a specific candidate.
-#[derive(Encode, Decode)]
-pub struct TrancheEntry {
-	pub tranche: DelayTranche,
-	// Assigned validators, and the instant we received their assignment, rounded
-	// to the nearest tick.
-	pub assignments: Vec<(ValidatorIndex, Tick)>,
-}
-
-/// Metadata regarding approval of a particular candidate within the context of some
-/// particular block.
-#[derive(Encode, Decode)]
-pub struct ApprovalEntry {
-	pub tranches: Vec<TrancheEntry>,
-	pub backing_group: GroupIndex,
-	pub our_assignment: Option<OurAssignment>,
-	// `n_validators` bits.
-	pub assignments: Bitfield,
-	pub approved: bool,
-}
-
-/// Metadata regarding approval of a particular candidate.
-#[derive(Encode, Decode)]
-pub struct CandidateEntry {
-	pub candidate: CandidateReceipt,
-	pub session: SessionIndex,
-	// Assignments are based on blocks, so we need to track assignments separately
-	// based on the block we are looking at.
-	pub block_assignments: BTreeMap<Hash, ApprovalEntry>,
-	pub approvals: Bitfield,
-}
-
-/// Metadata regarding approval of a particular block, by way of approval of the
-/// candidates contained within it.
-#[derive(Encode, Decode)]
-pub struct BlockEntry {
-	pub block_hash: Hash,
-	pub session: SessionIndex,
-	pub slot: Slot,
-	/// Random bytes derived from the VRF submitted within the block by the block
-	/// author as a credential and used as input to approval assignment criteria.
-	pub relay_vrf_story: [u8; 32],
-	// The candidates included as-of this block and the index of the core they are
-	// leaving. Sorted ascending by core index.
-	pub candidates: Vec<(CoreIndex, CandidateHash)>,
-	// A bitfield where the i'th bit corresponds to the i'th candidate in `candidates`.
-	// The i'th bit is `true` iff the candidate has been approved in the context of this
-	// block. The block can be considered approved if the bitfield has all bits set to `true`.
-	pub approved_bitfield: Bitfield,
-	pub children: Vec<Hash>,
-}
-
-/// A range from earliest..last block number stored within the DB.
-#[derive(Encode, Decode)]
-pub struct StoredBlockRange(BlockNumber, BlockNumber);
-
-impl From<crate::Tick> for Tick {
-	fn from(tick: crate::Tick) -> Tick {
-		tick
-	}
-}
-
-impl From<Tick> for crate::Tick {
-	fn from(tick: Tick) -> crate::Tick {
-		tick
-	}
-}
