+++ node/core/approval-voting/src/approval_db/v1/tests.rs
@@ -62,7 +62,7 @@ fn make_block_entry(
 		block_hash,
 		session: 1,
 		slot: Slot::from(1),
-		relay_vrf_story: RelayVRFStory([0u8; 32]),
+		relay_vrf_story: [0u8; 32],
 		approved_bitfield: make_bitvec(candidates.len()),
 		candidates,
 		children: Vec::new(),
